コンパイル警告や未使用の関数のコメントアウト
git-svn-id: http://dev.onionsoft.net/svn/openhsp/trunk@165 ceb3ebad-0e53-d446-8596-a45b2b8d79cd

+++ tools/win32/hsed3_footy2/Popfont.cpp
@@ -43,7 +43,7 @@ DWORD PopFontChooseColor(HWND hwnd, COLORREF crInitColor)
 	 if (res) {
 		return (DWORD)cc.rgbResult;
 	 }
-	 return -1;
+	 return (DWORD)-1;
 }
